third_party/blink/web_tests/html/tabular_data/tr_insertcell_default_argument-expected.txt

Tests that HTMLTableRowElement.insertCell()'s default argument is -1.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS tr.__proto__ is HTMLTableRowElement.prototype
PASS cells.length is 4
PASS cells[0].innerHTML is "0"
PASS cells[1].innerHTML is "1"
PASS cells[2].innerHTML is "2"
PASS cells[3].innerHTML is "3"
PASS successfullyParsed is true

TEST COMPLETE
